+++ src/main/java/com/xidian/reservation/dao/ConsumerMapper.java
@@ -3,13 +3,13 @@ package com.xidian.reservation.dao;
 import com.xidian.reservation.entity.Consumer;
 
 public interface ConsumerMapper {
-    int deleteByPrimaryKey(Integer consumerId);
+    int deleteByPrimaryKey(Long consumerId);
 
     int insert(Consumer record);
 
     int insertSelective(Consumer record);
 
-    Consumer selectByPrimaryKey(Integer consumerId);
+    Consumer selectByPrimaryKey(Long consumerId);
 
     int updateByPrimaryKeySelective(Consumer record);
